requirements.txt
aiofiles
bencode2
httpx
